backend/requirements.txt
streamlit
huggingface_hub
pillow
python-dotenv
